target_amount数据格式转换

+++ app/src/main/java/com/xiaomaoqiu/now/bussiness/bean/PetSportBean.java
@@ -15,7 +15,7 @@ public class PetSportBean extends BaseBean {
 
     public static class SportBean {
         public String date;
-        public int target_amount;
+        public double target_amount;
         public int reality_amount;
         public double percentage;
     }
